move files into Components and pages dsirectories
