+++ src/components/CreateUsername.tsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect, useContext } from "react";
-import { supabase } from "./supabaseClient";
+import { supabase } from ".././supabaseClient";
 import { UserOutlined } from "@ant-design/icons";
 import { Input, Button, Space } from "antd";
-import { AuthContext } from "./App";
+import { AuthContext } from ".././App";
 
 export const CreateUsername: React.FC<{}> = () => {
 	const [newUsername, setNewUsername] = useState<string>("");
