+++ src/components/WorkoutButton.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Button } from "antd";
-import { IWorkout } from "./WorkoutsPage";
-import { workouts } from "./data";
+import { IWorkout } from "../pages/WorkoutsPage";
+import { workouts } from "../data";
 
 const WorkoutButton: React.FC<{ workout: IWorkout }> = ({ workout }) => (
 	<Button type="primary" block className="workout-button">
